reorganization of map files
