+++ src/agdb/collections.rs
@@ -1,7 +1,7 @@
 pub mod bit_set;
 pub mod dictionary;
 pub mod map_storage;
-pub mod multi_map;
+pub mod old_multi_map;
 pub mod old_storage_vec;
 pub mod storage_dictionary;
 pub mod storage_map;
@@ -10,4 +10,5 @@ pub mod vec;
 pub mod vec_storage;
 
 mod map;
-mod map_common;
+mod old_map;
+mod old_map_common;
