+++ src/agdb/collections/old_map_common.rs
