+++ src/agdb/collections/old_map_common/map_data_storage.rs
