+++ src/agdb/collections/old_map_common/map_iterator.rs
