+++ src/agdb/collections/old_map_common/map_value_state.rs
